cohorts/2024/09-serverless/requirements.txt
numpy==1.24.3
pillow==10.4.0
requests==2.32.3
tensorflow==2.13.1
